fs: serve: Distribute requests only to idle workers
The original round-robin job queue distribution can cause starvation
when a request is distributed to a worker which has a long backlog of
slow requests to process. This is especially prevalent when there are a
lot of concurrent read and write requests, as mentioned in issue
chrislusf/seaweedfs#2263.

This commit uses `reflect.Select()` so that the request is only ever
sent to a channel in the ready state. As part of this change, all the
channels are now unbuffered instead of buffered, and the main FUSE
serving loop will block when no idle worker is available. There is no
point building up a job queue when no goroutine can process them -- it
will block the filesystem operation anyway, and increase memory usage
for no benefit. Just do not accept any new request when it cannot be
processed in time.

I am not sure if using the dynamic `select` provided by `reflect` is
faster than just creating a goroutine for each request, but doing this
at least allows us to still limit the maximum number of concurrent
requests to a reasonable level.

I'd also like to propose a customizable worker pool size `N` (and expose
it to the SeaweedFS side), but this is a topic for another day.